Project 2.0/requirements.txt
edgartools
chromadb
duckdb
sentence-transformers
google-genai
openai
rich
python-dotenv
langchain-text-splitters